requirements_led_server.txt
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
mss==10.0.0
numpy==2.2.1
paramiko==3.5.0
pillow==11.0.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
